feat: error handling

+++ src/storage/groups.ts
@@ -1,3 +1,4 @@
+import { GroupNameInUseError, UnexpectedError } from '@utils'
 import { GROUP_COLLECTION } from './storageConfig'
 
 import AsyncStorage from '@react-native-async-storage/async-storage'
@@ -6,19 +7,37 @@ export class GroupsStorage {
   private storageKey = GROUP_COLLECTION
 
   async create(name: string) {
-    const { groups } = await this.gelAll()
+    try {
+      const { groups } = await this.gelAll()
 
-    await AsyncStorage.setItem(
-      this.storageKey,
-      JSON.stringify([name, ...groups]),
-    )
+      const groupAlreadyExists = groups.includes(name)
+
+      if (groupAlreadyExists) throw new GroupNameInUseError()
+
+      await AsyncStorage.setItem(
+        this.storageKey,
+        JSON.stringify([name, ...groups]),
+      )
+    } catch (error) {
+      if (error instanceof GroupNameInUseError) {
+        throw error
+      } else {
+        console.error(error)
+        throw new UnexpectedError()
+      }
+    }
   }
 
   async gelAll() {
-    const storage = await AsyncStorage.getItem(this.storageKey)
+    try {
+      const storage = await AsyncStorage.getItem(this.storageKey)
 
-    const groups: string[] = storage ? JSON.parse(storage) : []
+      const groups: string[] = storage ? JSON.parse(storage) : []
 
-    return { groups }
+      return { groups }
+    } catch (error) {
+      console.error(error)
+      throw new UnexpectedError()
+    }
   }
 }
